# [`Optimzie`]: remove `AsParallel` in `Cache`