@Wi1l-B0t 2025-03-23
> It will depend of the amount to remove, isn't it?

Parallel sort is faster only when lots of data.
But there's only a small amount of data(all of `ECPointCache`, `ECDsaCache` and `RelayCache`)