@cschuchardt88 2025-03-23
It cost to much for overhead to setup parallel threading. Depends on the CPU count and number of items mostly.