@Jim8y 2025-03-24
| Method                          | OperationCount | Mean        | Error     | StdDev     | Median      |
|-------------------------------- |--------------- |------------:|----------:|-----------:|------------:|
| FIFOCacheAdd                    | 1000           |   854.18 us | 16.493 us |  17.647 us |   849.84 us |
| FIFOCacheContains               | 1000           |    19.37 us |  0.307 us |   0.257 us |    19.29 us |
| CachePerformanceWithoutParallel | 1000           |    74.96 us |  1.396 us |   2.002 us |    75.36 us |
| CachePerformanceWithParallel    | 1000           |   728.40 us | 34.200 us | 100.839 us |   701.56 us |
| FIFOCacheAdd                    | 10000          |   992.87 us |  2.781 us |   2.601 us |   992.61 us |
| FIFOCacheContains               | 10000          |    39.63 us |  0.115 us |   0.108 us |    39.67 us |
| CachePerformanceWithoutParallel | 10000          |   924.43 us |  4.093 us |   3.829 us |   925.26 us |
| CachePerformanceWithParallel    | 10000          | 2,937.06 us | 15.392 us |  12.853 us | 2,936.75 us |