@Jim8y 2025-03-24
well, from what i can see, AsParallel indeed is unnecessary.